Src/requirements.txt
pygame==2.0.1
pyopengl==3.1.5
numpy==1.19.5